requirements.txt
django==2.0.5
pillow >= 4.1, < 4.2
djangorestframework
django-rest-registration==0.3.9
imageio
numpy
djoser
django-bootstrap3
scikit-image
caffe
